+++ CHANGES.rst
@@ -11,6 +11,7 @@ in progress
 - Modernize to Python 3
 - Add compatibility with Grafana 6, 7 and 8
 - Add possibility to login to protected Grafana instances
+- Add possibility to control the navigation flavor (window, expand)
 
 
 2019-02-04 0.5.5